Setup
Read the array of strings which contains 2 elements: the first element will represent a list of comma-separated numbers sorted in ascending order., the second element is the same as the one element.

Goal - to return a string of comma-separated numbers which number that occur in both the first and second element. If no occurence happens then return string false.